requirements.txt
numpy
pandas
scikit-learn
joblib
Flask
mysql-connector-python
